Move FormatSize from attachment to output package
