+++ .go-arch-lint.yml
@@ -111,7 +111,6 @@ deps:
   # CLI layer — orchestrates everything
   cli:
     mayDependOn:
-      - attachment
       - dataentryconfig
       - filter
       - graph
