+++ internal/attachment/store.go
@@ -359,26 +359,6 @@ func detectContentType(ext string) string {
 	}
 }
 
-// FormatSize formats a byte size as a human-readable string.
-func FormatSize(size int64) string {
-	const (
-		KB = 1024
-		MB = KB * 1024
-		GB = MB * 1024
-	)
-
-	switch {
-	case size >= GB:
-		return fmt.Sprintf("%.1fGB", float64(size)/GB)
-	case size >= MB:
-		return fmt.Sprintf("%.1fMB", float64(size)/MB)
-	case size >= KB:
-		return fmt.Sprintf("%.1fKB", float64(size)/KB)
-	default:
-		return fmt.Sprintf("%dB", size)
-	}
-}
-
 // ValidatePath checks if a path is a valid attachment path.
 func ValidatePath(path string) error {
 	hash, ext, ok := ParsePath(path)
