+++ internal/output/output_test.go
@@ -635,6 +635,29 @@ func TestColorizeStatus(t *testing.T) {
 	}
 }
 
+// TestFormatSize tests byte size formatting
+func TestFormatSize(t *testing.T) {
+	tests := []struct {
+		size int64
+		want string
+	}{
+		{0, "0B"},
+		{100, "100B"},
+		{1024, "1.0KB"},
+		{1536, "1.5KB"},
+		{1048576, "1.0MB"},
+		{1572864, "1.5MB"},
+		{1073741824, "1.0GB"},
+	}
+
+	for _, tt := range tests {
+		got := FormatSize(tt.size)
+		if got != tt.want {
+			t.Errorf("FormatSize(%d) = %q, want %q", tt.size, got, tt.want)
+		}
+	}
+}
+
 // TestTruncate tests string truncation
 func TestTruncate(t *testing.T) {
 	tests := []struct {
